--- Navigation/Assets.xcassets/Stark/stark3.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "stark3.jpeg",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
