+++ Navigation/Assets.xcassets/joker.imageset/Contents.json
