+++ Navigation/Assets.xcassets/mega.imageset/Contents.json
