+++ Navigation/Assets.xcassets/mega.imageset/Mega.jpeg
